f17: Split up command.py to Model and Controller
